requirements.txt
Django==4.0
python-dotenv==0.20.0
psycopg2
social-auth-app-django
crispy-bootstrap5
celery
redis
django-celery-results
django-celery-beat
djangorestframework
drf-yasg